feat(map): avoid allocations in conversion from vector or slice (#16)
Most notably, introduce VecMap::from_vec_unchecked method which takes
ownership of a vector and constructs VecMap out of it without checking
for duplicate keys.  If client knows there are no duplicate keys in
the argument, they can use this to avoid quadratic checking done
otherwise.

With that, change `From<Vec>` implementation so that it doesn’t allocate
a new vector.  Instead, it deduplicates elements in provided vector and
then uses that storage for the constructed VecMap.

Lastly, add FromIterator and Extend implementations for `&(K, V)` items
so that conversion from slice doesn’t have to implement temporary vector.

Co-authored-by: martinohmann <martinohmann@gmail.com>